requirements.txt
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioice==0.10.2
aiortc==1.14.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
av==16.1.0
babel==2.18.0
blis==1.3.3
brotli==1.2.0
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
coloredlogs==15.0.1
confection==0.1.5
cryptography==46.0.4
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cymem==2.0.13
dacite==1.9.2
datasets==4.5.0
decorator==5.2.1
dill==0.4.0
dlinfo==2.0.0
dnspython==2.8.0
docopt==0.6.2
einops==0.8.2
einx==0.3.0
espeakng-loader==0.2.4
fastapi==0.128.0
fastrtc==0.0.34
fastrtc-moonshine-onnx==20241016
ffmpy==1.0.0
filelock==3.20.3
flatbuffers==25.12.19
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.1.0
google-crc32c==1.8.0
gradio==5.50.0
gradio-client==1.14.0
groovy==0.1.2
h11==0.16.0
hf-transfer==0.1.9
hf-xet==1.3.0b0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.3.5
humanfriendly==10.0
idna==3.11
ifaddr==0.2.0
isodate==0.7.2
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
language-tags==1.2.0
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.46.0
loguru==0.7.3
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
miniaudio==1.61
misaki==0.9.4
mistral-common==1.9.0
mlx==0.30.4
mlx-audio==0.3.1
mlx-lm==0.30.5
mlx-metal==0.30.4
mlx-vlm==0.3.9
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
networkx==3.6.1
num2words==0.5.14
numba==0.63.0b1
numpy==2.4.2
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python==4.13.0.90
orjson==3.11.6
packaging==26.0
pandas==2.3.3
phonemizer-fork==3.3.2
pillow==11.3.0
pip==26.0
platformdirs==4.5.1
pooch==1.9.0
preshed==3.0.12
propcache==0.4.1
protobuf==7.34.0rc1
psutil==7.2.2
pyarrow==23.0.0
pycountry==24.6.1
pycparser==3.0
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-extra-types==2.11.0
pydub==0.25.1
pyee==13.0.0
pygments==2.19.2
pylibsrtp==1.0.0
pyloudnorm==0.2.0
pyopenssl==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
pyyaml==6.0.3
rdflib==7.5.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rfc3986==1.5.0
rich==14.3.2
rpds-py==0.30.0
ruff==0.14.14
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
segments==2.3.0
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smart-open==7.5.0
sounddevice==0.5.3
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
spacy==3.8.11
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.50.0
sympy==1.14.0
termcolor==3.3.0
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomlkit==0.13.3
tqdm==4.67.2
transformers==5.0.0rc3
typer==0.21.1
typer-slim==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
wasabi==1.1.3
weasel==0.4.3
webrtcvad==2.0.10
websockets==15.0.1
wrapt==2.1.0
xxhash==3.6.0
yarl==1.22.0
